backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0
PyMuPDF==1.24.12
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
